requirements.txt
tensorflow
gradio
numpy
pillow
huggingface_hub
